Make namespace types splitting the default
