+++ Sources/swift-openapi-generator/Documentation.docc/Articles/Configuring-the-generator.md
@@ -48,12 +48,6 @@ The configuration file has the following keys:
 - `typeOverrides` (optional): Allows replacing a generated type with a custom type.
     - `schemas` (optional): a string to string dictionary. The key is the name of the schema, the last component of `#/components/schemas/Foo` (here, `Foo`). The value is the custom type name, such as `CustomFoo`. Check out details in [SOAR-0014](https://swiftpackageindex.com/apple/swift-openapi-generator/documentation/swift-openapi-generator/soar-0014).
 - `featureFlags` (optional): array of strings. Each string must be a valid feature flag to enable. For a list of currently supported feature flags, check out [FeatureFlags.swift](https://github.com/apple/swift-openapi-generator/blob/main/Sources/_OpenAPIGeneratorCore/FeatureFlags.swift).
-- `output` (optional): Options that affect generated output files.
-    - `types` (optional): Options that only affect `types` generation.
-        - `fileSplitting` (optional): Splits generated types across multiple files.
-            - `strategy` (required): The file splitting strategy. Known values:
-                - `namespace`: Split generated types into a small fixed set of files by top-level namespace.
-            - `namespace` (optional): Options for the `namespace` strategy.
 
 ### Example config files
 
@@ -116,21 +110,20 @@ additionalFileComments:
   - "swiftlint:disable all"
 ```
 
-To split generated types across multiple Swift files:
+Types generation emits a fixed set of files organized by generated namespace:
 
-```yaml
-generate:
-  - types
-namingStrategy: idiomatic
-output:
-  types:
-    fileSplitting:
-      strategy: namespace
-```
+- `Types.swift`
+- `Types+Components.swift`
+- `Types+Operations.swift`
+- `Types+Components+Schemas.swift`
+- `Types+Components+Parameters.swift`
+- `Types+Components+RequestBodies.swift`
+- `Types+Components+Responses.swift`
+- `Types+Components+Headers.swift`
 
-The same strategy can be enabled from the command line with `--types-file-splitting namespace`.
-
-> Note: Types file splitting is currently supported when running the generator directly or with the command plugin. The build tool plugin does not yet support types file splitting.
+The file layout does not change generated Swift symbol names. For example, schema types remain nested under
+`Components.Schemas`. When invoking the generator directly or checking generated sources into a repository, retain all
+of the emitted files.
 
 ### Document filtering
