+++ Sources/swift-openapi-generator/GenerateOptions+runGenerator.swift
@@ -37,12 +37,6 @@ extension _GenerateOptions {
         let resolvedNameOverrides = resolvedNameOverrides(config)
         let resolvedTypeOverrides = resolvedTypeOverrides(config)
         let resolvedFeatureFlags = resolvedFeatureFlags(config)
-        let resolvedOutputOptions = resolvedOutputOptions(config)
-        guard pluginSource != .build || resolvedOutputOptions.types?.fileSplitting == nil else {
-            throw ValidationError(
-                "Types file splitting is not supported by the build tool plugin yet. Run swift-openapi-generator directly or use the command plugin to generate split files."
-            )
-        }
         let configs: [Config] = sortedModes.map {
             .init(
                 mode: $0,
@@ -53,8 +47,7 @@ extension _GenerateOptions {
                 namingStrategy: resolvedNamingStragy,
                 nameOverrides: resolvedNameOverrides,
                 typeOverrides: resolvedTypeOverrides,
-                featureFlags: resolvedFeatureFlags,
-                output: resolvedOutputOptions
+                featureFlags: resolvedFeatureFlags
             )
         }
         let (diagnostics, finalizeDiagnostics) = preparedDiagnosticsCollector(outputPath: diagnosticsOutputPath)
@@ -74,8 +67,7 @@ extension _GenerateOptions {
                 .sorted(by: { $0.key < $1.key })
                 .map { "\"\($0.key)\"->\"\($0.value)\"" }.joined(separator: ", "))
             - Feature flags: \(resolvedFeatureFlags.isEmpty ? "<none>" : resolvedFeatureFlags.map(\.rawValue).joined(separator: ", "))
-            - Types file splitting: \(resolvedOutputOptions.types?.fileSplitting?.strategy.rawValue ?? "<none>")
-            - Output file names: \(sortedModes.map(\.outputFileName).joined(separator: ", "))
+            - Output file names: \(sortedModes.flatMap(\.outputFileNames).joined(separator: ", "))
             - Output directory: \(outputDirectory.path)
             - Diagnostics output path: \(diagnosticsOutputPath?.path ?? "<none - logs to stderr>")
             - Current directory: \(FileManager.default.currentDirectoryPath)
