+++ Sources/swift-openapi-generator/UserConfig.swift
@@ -51,9 +51,6 @@ struct _UserConfig: Codable {
     /// A set of features to explicitly enable.
     var featureFlags: FeatureFlags?
 
-    /// Options that affect generated output files.
-    var output: OutputOptions?
-
     /// A set of raw values corresponding to the coding keys of this struct.
     static let codingKeysRawValues = Set(CodingKeys.allCases.map({ $0.rawValue }))
 
@@ -67,7 +64,6 @@ struct _UserConfig: Codable {
         case nameOverrides
         case typeOverrides
         case featureFlags
-        case output
     }
 
     /// A container of type overrides.
