+++ Tests/OpenAPIGeneratorCoreTests/Test_Config.swift
@@ -31,33 +31,28 @@ final class Test_Config: Test_Core {
         XCTAssertEqual(config.additionalFileComments, [])
     }
 
-    func testOutputOptionsDefaultToEmpty() {
-        let config = Config(mode: .types, access: .public, namingStrategy: .defensive)
-        XCTAssertNil(config.output.types)
-    }
-
-    func testTypesFileSplittingConfig() {
-        let config = Config(
-            mode: .types,
-            access: .public,
-            namingStrategy: .defensive,
-            output: .init(types: .init(fileSplitting: .init(strategy: .namespace)))
-        )
-        XCTAssertEqual(config.output.types?.fileSplitting?.strategy, .namespace)
-    }
-
     func testGeneratorModeOutputFileNameHelper() {
         XCTAssertEqual(GeneratorMode.outputFileName("Types"), "Types.swift")
         XCTAssertEqual(GeneratorMode.outputFileName("Types.swift", "Components"), "Types+Components.swift")
         XCTAssertEqual(GeneratorMode.outputFileName("Types.swift", "Operations.swift"), "Types+Operations.swift")
     }
 
-    func testNamespaceFileSplittingOutputFileNames() {
-        let config = TypesFileSplittingConfig(strategy: .namespace)
-
+    func testGeneratorModeOutputFileNames() {
         XCTAssertEqual(
-            config.outputFileNames(primaryTypesFileName: "Types.swift"),
-            ["Types.swift", "Types+Components.swift", "Types+Operations.swift"]
+            GeneratorMode.types.outputFileNames,
+            [
+                "Types.swift",
+                "Types+Components.swift",
+                "Types+Operations.swift",
+                "Types+Components+Schemas.swift",
+                "Types+Components+Parameters.swift",
+                "Types+Components+RequestBodies.swift",
+                "Types+Components+Responses.swift",
+                "Types+Components+Headers.swift",
+            ]
         )
+        XCTAssertEqual(GeneratorMode.client.outputFileNames, ["Client.swift"])
+        XCTAssertEqual(GeneratorMode.server.outputFileNames, ["Server.swift"])
+        XCTAssertEqual(GeneratorMode.allOutputFileNames, GeneratorMode.allCases.flatMap(\.outputFileNames))
     }
 }
