+++ Tests/OpenAPIGeneratorReferenceTests/Resources/ReferenceSources/Petstore/Types+Components+Headers.swift
@@ -1,0 +1,20 @@
+// Generated by swift-openapi-generator, do not modify.
+@_spi(Generated) import OpenAPIRuntime
+#if os(Linux)
+@preconcurrency import struct Foundation.URL
+@preconcurrency import struct Foundation.Data
+@preconcurrency import struct Foundation.Date
+#else
+import struct Foundation.URL
+import struct Foundation.Data
+import struct Foundation.Date
+#endif
+extension Components {
+    /// Types generated from the `#/components/headers` section of the OpenAPI document.
+    public enum Headers {
+        /// A description here.
+        ///
+        /// - Remark: Generated from `#/components/headers/TracingHeader`.
+        public typealias TracingHeader = Swift.String
+    }
+}
