+++ Tests/OpenAPIGeneratorReferenceTests/Resources/ReferenceSources/Petstore/Types+Components+Parameters.swift
@@ -1,0 +1,24 @@
+// Generated by swift-openapi-generator, do not modify.
+@_spi(Generated) import OpenAPIRuntime
+#if os(Linux)
+@preconcurrency import struct Foundation.URL
+@preconcurrency import struct Foundation.Data
+@preconcurrency import struct Foundation.Date
+#else
+import struct Foundation.URL
+import struct Foundation.Data
+import struct Foundation.Date
+#endif
+extension Components {
+    /// Types generated from the `#/components/parameters` section of the OpenAPI document.
+    public enum Parameters {
+        /// Supply this parameter to filter pets born since the provided date.
+        ///
+        /// - Remark: Generated from `#/components/parameters/query.born-since`.
+        public typealias Query_bornSince = Components.Schemas.Dob
+        /// The id of the pet to retrieve
+        ///
+        /// - Remark: Generated from `#/components/parameters/path.petId`.
+        public typealias Path_petId = Swift.Int64
+    }
+}
