+++ Tests/OpenAPIGeneratorReferenceTests/Resources/ReferenceSources/Petstore/Types+Components+RequestBodies.swift
@@ -1,0 +1,138 @@
+// Generated by swift-openapi-generator, do not modify.
+@_spi(Generated) import OpenAPIRuntime
+#if os(Linux)
+@preconcurrency import struct Foundation.URL
+@preconcurrency import struct Foundation.Data
+@preconcurrency import struct Foundation.Date
+#else
+import struct Foundation.URL
+import struct Foundation.Data
+import struct Foundation.Date
+#endif
+extension Components {
+    /// Types generated from the `#/components/requestBodies` section of the OpenAPI document.
+    public enum RequestBodies {
+        /// - Remark: Generated from `#/components/requestBodies/UpdatePetRequest`.
+        @frozen public enum UpdatePetRequest: Sendable, Hashable {
+            /// - Remark: Generated from `#/components/requestBodies/UpdatePetRequest/json`.
+            public struct JsonPayload: Codable, Hashable, Sendable {
+                /// Pet name
+                ///
+                /// - Remark: Generated from `#/components/requestBodies/UpdatePetRequest/json/name`.
+                public var name: Swift.String?
+                /// - Remark: Generated from `#/components/requestBodies/UpdatePetRequest/json/kind`.
+                public var kind: Components.Schemas.PetKind?
+                /// - Remark: Generated from `#/components/requestBodies/UpdatePetRequest/json/tag`.
+                public var tag: Swift.String?
+                /// Creates a new `JsonPayload`.
+                ///
+                /// - Parameters:
+                ///   - name: Pet name
+                ///   - kind:
+                ///   - tag:
+                public init(
+                    name: Swift.String? = nil,
+                    kind: Components.Schemas.PetKind? = nil,
+                    tag: Swift.String? = nil
+                ) {
+                    self.name = name
+                    self.kind = kind
+                    self.tag = tag
+                }
+                public enum CodingKeys: String, CodingKey {
+                    case name
+                    case kind
+                    case tag
+                }
+            }
+            /// - Remark: Generated from `#/components/requestBodies/UpdatePetRequest/content/application\/json`.
+            case json(Components.RequestBodies.UpdatePetRequest.JsonPayload)
+        }
+        /// - Remark: Generated from `#/components/requestBodies/MultipartUploadTypedRequest`.
+        @frozen public enum MultipartUploadTypedRequest: Sendable, Hashable {
+            /// - Remark: Generated from `#/components/requestBodies/MultipartUploadTypedRequest/multipartForm`.
+            @frozen public enum MultipartFormPayload: Sendable, Hashable {
+                /// - Remark: Generated from `#/components/requestBodies/MultipartUploadTypedRequest/multipartForm/log`.
+                public struct LogPayload: Sendable, Hashable {
+                    /// - Remark: Generated from `#/components/requestBodies/MultipartUploadTypedRequest/multipartForm/log/headers`.
+                    public struct Headers: Sendable, Hashable {
+                        /// - Remark: Generated from `#/components/requestBodies/MultipartUploadTypedRequest/multipartForm/log/headers/x-log-type`.
+                        @frozen public enum XLogTypePayload: String, Codable, Hashable, Sendable, CaseIterable {
+                            case structured = "structured"
+                            case unstructured = "unstructured"
+                        }
+                        /// The type of the log.
+                        ///
+                        /// - Remark: Generated from `#/components/requestBodies/MultipartUploadTypedRequest/multipartForm/log/headers/x-log-type`.
+                        public var xLogType: Components.RequestBodies.MultipartUploadTypedRequest.MultipartFormPayload.LogPayload.Headers.XLogTypePayload?
+                        /// Creates a new `Headers`.
+                        ///
+                        /// - Parameters:
+                        ///   - xLogType: The type of the log.
+                        public init(xLogType: Components.RequestBodies.MultipartUploadTypedRequest.MultipartFormPayload.LogPayload.Headers.XLogTypePayload? = nil) {
+                            self.xLogType = xLogType
+                        }
+                    }
+                    /// Received HTTP response headers
+                    public var headers: Components.RequestBodies.MultipartUploadTypedRequest.MultipartFormPayload.LogPayload.Headers
+                    public var body: OpenAPIRuntime.HTTPBody
+                    /// Creates a new `LogPayload`.
+                    ///
+                    /// - Parameters:
+                    ///   - headers: Received HTTP response headers
+                    ///   - body:
+                    public init(
+                        headers: Components.RequestBodies.MultipartUploadTypedRequest.MultipartFormPayload.LogPayload.Headers = .init(),
+                        body: OpenAPIRuntime.HTTPBody
+                    ) {
+                        self.headers = headers
+                        self.body = body
+                    }
+                }
+                case log(OpenAPIRuntime.MultipartPart<Components.RequestBodies.MultipartUploadTypedRequest.MultipartFormPayload.LogPayload>)
+                /// - Remark: Generated from `#/components/requestBodies/MultipartUploadTypedRequest/multipartForm/metadata`.
+                public struct MetadataPayload: Sendable, Hashable {
+                    /// - Remark: Generated from `#/components/requestBodies/MultipartUploadTypedRequest/multipartForm/metadata/content/body`.
+                    public struct BodyPayload: Codable, Hashable, Sendable {
+                        /// - Remark: Generated from `#/components/requestBodies/MultipartUploadTypedRequest/multipartForm/metadata/content/body/createdAt`.
+                        public var createdAt: Foundation.Date
+                        /// Creates a new `BodyPayload`.
+                        ///
+                        /// - Parameters:
+                        ///   - createdAt:
+                        public init(createdAt: Foundation.Date) {
+                            self.createdAt = createdAt
+                        }
+                        public enum CodingKeys: String, CodingKey {
+                            case createdAt
+                        }
+                    }
+                    public var body: Components.RequestBodies.MultipartUploadTypedRequest.MultipartFormPayload.MetadataPayload.BodyPayload
+                    /// Creates a new `MetadataPayload`.
+                    ///
+                    /// - Parameters:
+                    ///   - body:
+                    public init(body: Components.RequestBodies.MultipartUploadTypedRequest.MultipartFormPayload.MetadataPayload.BodyPayload) {
+                        self.body = body
+                    }
+                }
+                case metadata(OpenAPIRuntime.MultipartPart<Components.RequestBodies.MultipartUploadTypedRequest.MultipartFormPayload.MetadataPayload>)
+                /// - Remark: Generated from `#/components/requestBodies/MultipartUploadTypedRequest/multipartForm/keyword`.
+                public struct KeywordPayload: Sendable, Hashable {
+                    public var body: OpenAPIRuntime.HTTPBody
+                    /// Creates a new `KeywordPayload`.
+                    ///
+                    /// - Parameters:
+                    ///   - body:
+                    public init(body: OpenAPIRuntime.HTTPBody) {
+                        self.body = body
+                    }
+                }
+                case keyword(OpenAPIRuntime.MultipartPart<Components.RequestBodies.MultipartUploadTypedRequest.MultipartFormPayload.KeywordPayload>)
+                case undocumented(OpenAPIRuntime.MultipartRawPart)
+            }
+            /// - Remark: Generated from `#/components/requestBodies/MultipartUploadTypedRequest/content/multipart\/form-data`.
+            case multipartForm(OpenAPIRuntime.MultipartBody<Components.RequestBodies.MultipartUploadTypedRequest.MultipartFormPayload>)
+        }
+    }
+}
