+++ Tests/OpenAPIGeneratorReferenceTests/Resources/ReferenceSources/Petstore/Types+Components+Responses.swift
@@ -1,0 +1,189 @@
+// Generated by swift-openapi-generator, do not modify.
+@_spi(Generated) import OpenAPIRuntime
+#if os(Linux)
+@preconcurrency import struct Foundation.URL
+@preconcurrency import struct Foundation.Data
+@preconcurrency import struct Foundation.Date
+#else
+import struct Foundation.URL
+import struct Foundation.Data
+import struct Foundation.Date
+#endif
+extension Components {
+    /// Types generated from the `#/components/responses` section of the OpenAPI document.
+    public enum Responses {
+        public struct ErrorBadRequest: Sendable, Hashable {
+            /// - Remark: Generated from `#/components/responses/ErrorBadRequest/headers`.
+            public struct Headers: Sendable, Hashable {
+                /// A description here.
+                ///
+                /// - Remark: Generated from `#/components/responses/ErrorBadRequest/headers/X-Reason`.
+                public var xReason: Swift.String?
+                /// Creates a new `Headers`.
+                ///
+                /// - Parameters:
+                ///   - xReason: A description here.
+                public init(xReason: Swift.String? = nil) {
+                    self.xReason = xReason
+                }
+            }
+            /// Received HTTP response headers
+            public var headers: Components.Responses.ErrorBadRequest.Headers
+            /// - Remark: Generated from `#/components/responses/ErrorBadRequest/content`.
+            @frozen public enum Body: Sendable, Hashable {
+                /// - Remark: Generated from `#/components/responses/ErrorBadRequest/content/json`.
+                public struct JsonPayload: Codable, Hashable, Sendable {
+                    /// - Remark: Generated from `#/components/responses/ErrorBadRequest/content/json/code`.
+                    public var code: Swift.Int
+                    /// Creates a new `JsonPayload`.
+                    ///
+                    /// - Parameters:
+                    ///   - code:
+                    public init(code: Swift.Int) {
+                        self.code = code
+                    }
+                    public enum CodingKeys: String, CodingKey {
+                        case code
+                    }
+                }
+                /// - Remark: Generated from `#/components/responses/ErrorBadRequest/content/application\/json`.
+                case json(Components.Responses.ErrorBadRequest.Body.JsonPayload)
+                /// The associated value of the enum case if `self` is `.json`.
+                ///
+                /// - Throws: An error if `self` is not `.json`.
+                /// - SeeAlso: `.json`.
+                public var json: Components.Responses.ErrorBadRequest.Body.JsonPayload {
+                    get throws {
+                        switch self {
+                        case let .json(body):
+                            return body
+                        }
+                    }
+                }
+            }
+            /// Received HTTP response body
+            public var body: Components.Responses.ErrorBadRequest.Body
+            /// Creates a new `ErrorBadRequest`.
+            ///
+            /// - Parameters:
+            ///   - headers: Received HTTP response headers
+            ///   - body: Received HTTP response body
+            public init(
+                headers: Components.Responses.ErrorBadRequest.Headers = .init(),
+                body: Components.Responses.ErrorBadRequest.Body
+            ) {
+                self.headers = headers
+                self.body = body
+            }
+        }
+        public struct MultipartDownloadTypedResponse: Sendable, Hashable {
+            /// - Remark: Generated from `#/components/responses/MultipartDownloadTypedResponse/content`.
+            @frozen public enum Body: Sendable, Hashable {
+                /// - Remark: Generated from `#/components/responses/MultipartDownloadTypedResponse/content/multipartForm`.
+                @frozen public enum MultipartFormPayload: Sendable, Hashable {
+                    /// - Remark: Generated from `#/components/responses/MultipartDownloadTypedResponse/content/multipartForm/log`.
+                    public struct LogPayload: Sendable, Hashable {
+                        /// - Remark: Generated from `#/components/responses/MultipartDownloadTypedResponse/content/multipartForm/log/headers`.
+                        public struct Headers: Sendable, Hashable {
+                            /// - Remark: Generated from `#/components/responses/MultipartDownloadTypedResponse/content/multipartForm/log/headers/x-log-type`.
+                            @frozen public enum XLogTypePayload: String, Codable, Hashable, Sendable, CaseIterable {
+                                case structured = "structured"
+                                case unstructured = "unstructured"
+                            }
+                            /// The type of the log.
+                            ///
+                            /// - Remark: Generated from `#/components/responses/MultipartDownloadTypedResponse/content/multipartForm/log/headers/x-log-type`.
+                            public var xLogType: Components.Responses.MultipartDownloadTypedResponse.Body.MultipartFormPayload.LogPayload.Headers.XLogTypePayload?
+                            /// Creates a new `Headers`.
+                            ///
+                            /// - Parameters:
+                            ///   - xLogType: The type of the log.
+                            public init(xLogType: Components.Responses.MultipartDownloadTypedResponse.Body.MultipartFormPayload.LogPayload.Headers.XLogTypePayload? = nil) {
+                                self.xLogType = xLogType
+                            }
+                        }
+                        /// Received HTTP response headers
+                        public var headers: Components.Responses.MultipartDownloadTypedResponse.Body.MultipartFormPayload.LogPayload.Headers
+                        public var body: OpenAPIRuntime.HTTPBody
+                        /// Creates a new `LogPayload`.
+                        ///
+                        /// - Parameters:
+                        ///   - headers: Received HTTP response headers
+                        ///   - body:
+                        public init(
+                            headers: Components.Responses.MultipartDownloadTypedResponse.Body.MultipartFormPayload.LogPayload.Headers = .init(),
+                            body: OpenAPIRuntime.HTTPBody
+                        ) {
+                            self.headers = headers
+                            self.body = body
+                        }
+                    }
+                    case log(OpenAPIRuntime.MultipartPart<Components.Responses.MultipartDownloadTypedResponse.Body.MultipartFormPayload.LogPayload>)
+                    /// - Remark: Generated from `#/components/responses/MultipartDownloadTypedResponse/content/multipartForm/metadata`.
+                    public struct MetadataPayload: Sendable, Hashable {
+                        /// - Remark: Generated from `#/components/responses/MultipartDownloadTypedResponse/content/multipartForm/metadata/content/body`.
+                        public struct BodyPayload: Codable, Hashable, Sendable {
+                            /// - Remark: Generated from `#/components/responses/MultipartDownloadTypedResponse/content/multipartForm/metadata/content/body/createdAt`.
+                            public var createdAt: Foundation.Date
+                            /// Creates a new `BodyPayload`.
+                            ///
+                            /// - Parameters:
+                            ///   - createdAt:
+                            public init(createdAt: Foundation.Date) {
+                                self.createdAt = createdAt
+                            }
+                            public enum CodingKeys: String, CodingKey {
+                                case createdAt
+                            }
+                        }
+                        public var body: Components.Responses.MultipartDownloadTypedResponse.Body.MultipartFormPayload.MetadataPayload.BodyPayload
+                        /// Creates a new `MetadataPayload`.
+                        ///
+                        /// - Parameters:
+                        ///   - body:
+                        public init(body: Components.Responses.MultipartDownloadTypedResponse.Body.MultipartFormPayload.MetadataPayload.BodyPayload) {
+                            self.body = body
+                        }
+                    }
+                    case metadata(OpenAPIRuntime.MultipartPart<Components.Responses.MultipartDownloadTypedResponse.Body.MultipartFormPayload.MetadataPayload>)
+                    /// - Remark: Generated from `#/components/responses/MultipartDownloadTypedResponse/content/multipartForm/keyword`.
+                    public struct KeywordPayload: Sendable, Hashable {
+                        public var body: OpenAPIRuntime.HTTPBody
+                        /// Creates a new `KeywordPayload`.
+                        ///
+                        /// - Parameters:
+                        ///   - body:
+                        public init(body: OpenAPIRuntime.HTTPBody) {
+                            self.body = body
+                        }
+                    }
+                    case keyword(OpenAPIRuntime.MultipartPart<Components.Responses.MultipartDownloadTypedResponse.Body.MultipartFormPayload.KeywordPayload>)
+                    case undocumented(OpenAPIRuntime.MultipartRawPart)
+                }
+                /// - Remark: Generated from `#/components/responses/MultipartDownloadTypedResponse/content/multipart\/form-data`.
+                case multipartForm(OpenAPIRuntime.MultipartBody<Components.Responses.MultipartDownloadTypedResponse.Body.MultipartFormPayload>)
+                /// The associated value of the enum case if `self` is `.multipartForm`.
+                ///
+                /// - Throws: An error if `self` is not `.multipartForm`.
+                /// - SeeAlso: `.multipartForm`.
+                public var multipartForm: OpenAPIRuntime.MultipartBody<Components.Responses.MultipartDownloadTypedResponse.Body.MultipartFormPayload> {
+                    get throws {
+                        switch self {
+                        case let .multipartForm(body):
+                            return body
+                        }
+                    }
+                }
+            }
+            /// Received HTTP response body
+            public var body: Components.Responses.MultipartDownloadTypedResponse.Body
+            /// Creates a new `MultipartDownloadTypedResponse`.
+            ///
+            /// - Parameters:
+            ///   - body: Received HTTP response body
+            public init(body: Components.Responses.MultipartDownloadTypedResponse.Body) {
+                self.body = body
+            }
+        }
+    }
+}
