+++ Tests/OpenAPIGeneratorReferenceTests/Resources/ReferenceSources/Petstore/Types+Components+Schemas.swift
@@ -1,0 +1,1277 @@
+// Generated by swift-openapi-generator, do not modify.
+@_spi(Generated) import OpenAPIRuntime
+#if os(Linux)
+@preconcurrency import struct Foundation.URL
+@preconcurrency import struct Foundation.Data
+@preconcurrency import struct Foundation.Date
+#else
+import struct Foundation.URL
+import struct Foundation.Data
+import struct Foundation.Date
+#endif
+extension Components {
+    /// Types generated from the `#/components/schemas` section of the OpenAPI document.
+    public enum Schemas {
+        /// Pet metadata
+        ///
+        /// - Remark: Generated from `#/components/schemas/Pet`.
+        public struct Pet: Codable, Hashable, Sendable {
+            /// Pet id
+            ///
+            /// - Remark: Generated from `#/components/schemas/Pet/id`.
+            public var id: Swift.Int64
+            /// Pet name
+            ///
+            /// - Remark: Generated from `#/components/schemas/Pet/name`.
+            public var name: Swift.String
+            /// - Remark: Generated from `#/components/schemas/Pet/tag`.
+            public var tag: Swift.String?
+            /// Pet genome (base64-encoded)
+            ///
+            /// - Remark: Generated from `#/components/schemas/Pet/genome`.
+            public var genome: OpenAPIRuntime.Base64EncodedData?
+            /// - Remark: Generated from `#/components/schemas/Pet/kind`.
+            public var kind: Components.Schemas.PetKind?
+            /// Creates a new `Pet`.
+            ///
+            /// - Parameters:
+            ///   - id: Pet id
+            ///   - name: Pet name
+            ///   - tag:
+            ///   - genome: Pet genome (base64-encoded)
+            ///   - kind:
+            public init(
+                id: Swift.Int64,
+                name: Swift.String,
+                tag: Swift.String? = nil,
+                genome: OpenAPIRuntime.Base64EncodedData? = nil,
+                kind: Components.Schemas.PetKind? = nil
+            ) {
+                self.id = id
+                self.name = name
+                self.tag = tag
+                self.genome = genome
+                self.kind = kind
+            }
+            public enum CodingKeys: String, CodingKey {
+                case id
+                case name
+                case tag
+                case genome
+                case kind
+            }
+        }
+        /// - Remark: Generated from `#/components/schemas/MixedAnyOf`.
+        public struct MixedAnyOf: Codable, Hashable, Sendable {
+            /// - Remark: Generated from `#/components/schemas/MixedAnyOf/value1`.
+            public var value1: Foundation.Date?
+            /// - Remark: Generated from `#/components/schemas/MixedAnyOf/value2`.
+            public var value2: Components.Schemas.PetKind?
+            /// - Remark: Generated from `#/components/schemas/MixedAnyOf/value3`.
+            public var value3: Components.Schemas.Pet?
+            /// - Remark: Generated from `#/components/schemas/MixedAnyOf/value4`.
+            public var value4: Swift.String?
+            /// Creates a new `MixedAnyOf`.
+            ///
+            /// - Parameters:
+            ///   - value1:
+            ///   - value2:
+            ///   - value3:
+            ///   - value4:
+            public init(
+                value1: Foundation.Date? = nil,
+                value2: Components.Schemas.PetKind? = nil,
+                value3: Components.Schemas.Pet? = nil,
+                value4: Swift.String? = nil
+            ) {
+                self.value1 = value1
+                self.value2 = value2
+                self.value3 = value3
+                self.value4 = value4
+            }
+            public init(from decoder: any Swift.Decoder) throws {
+                var errors: [any Swift.Error] = []
+                do {
+                    self.value1 = try decoder.decodeFromSingleValueContainer()
+                } catch {
+                    errors.append(error)
+                }
+                do {
+                    self.value2 = try decoder.decodeFromSingleValueContainer()
+                } catch {
+                    errors.append(error)
+                }
+                do {
+                    self.value3 = try .init(from: decoder)
+                } catch {
+                    errors.append(error)
+                }
+                do {
+                    self.value4 = try decoder.decodeFromSingleValueContainer()
+                } catch {
+                    errors.append(error)
+                }
+                try Swift.DecodingError.verifyAtLeastOneSchemaIsNotNil(
+                    [
+                        self.value1,
+                        self.value2,
+                        self.value3,
+                        self.value4
+                    ],
+                    type: Self.self,
+                    codingPath: decoder.codingPath,
+                    errors: errors
+                )
+            }
+            public func encode(to encoder: any Swift.Encoder) throws {
+                try encoder.encodeFirstNonNilValueToSingleValueContainer([
+                    self.value1,
+                    self.value2,
+                    self.value4
+                ])
+                try self.value3?.encode(to: encoder)
+            }
+        }
+        /// - Remark: Generated from `#/components/schemas/MixedOneOf`.
+        @frozen public enum MixedOneOf: Codable, Hashable, Sendable {
+            /// - Remark: Generated from `#/components/schemas/MixedOneOf/case1`.
+            case case1(Foundation.Date)
+            /// - Remark: Generated from `#/components/schemas/MixedOneOf/case2`.
+            case PetKind(Components.Schemas.PetKind)
+            /// - Remark: Generated from `#/components/schemas/MixedOneOf/case3`.
+            case Pet(Components.Schemas.Pet)
+            public init(from decoder: any Swift.Decoder) throws {
+                var errors: [any Swift.Error] = []
+                do {
+                    self = .case1(try decoder.decodeFromSingleValueContainer())
+                    return
+                } catch {
+                    errors.append(error)
+                }
+                do {
+                    self = .PetKind(try decoder.decodeFromSingleValueContainer())
+                    return
+                } catch {
+                    errors.append(error)
+                }
+                do {
+                    self = .Pet(try .init(from: decoder))
+                    return
+                } catch {
+                    errors.append(error)
+                }
+                throw Swift.DecodingError.failedToDecodeOneOfSchema(
+                    type: Self.self,
+                    codingPath: decoder.codingPath,
+                    errors: errors
+                )
+            }
+            public func encode(to encoder: any Swift.Encoder) throws {
+                switch self {
+                case let .case1(value):
+                    try encoder.encodeToSingleValueContainer(value)
+                case let .PetKind(value):
+                    try encoder.encodeToSingleValueContainer(value)
+                case let .Pet(value):
+                    try value.encode(to: encoder)
+                }
+            }
+        }
+        /// - Remark: Generated from `#/components/schemas/MixedAllOfPrimitive`.
+        public struct MixedAllOfPrimitive: Codable, Hashable, Sendable {
+            /// - Remark: Generated from `#/components/schemas/MixedAllOfPrimitive/value1`.
+            public var value1: Foundation.Date
+            /// - Remark: Generated from `#/components/schemas/MixedAllOfPrimitive/value2`.
+            public var value2: Swift.String
+            /// Creates a new `MixedAllOfPrimitive`.
+            ///
+            /// - Parameters:
+            ///   - value1:
+            ///   - value2:
+            public init(
+                value1: Foundation.Date,
+                value2: Swift.String
+            ) {
+                self.value1 = value1
+                self.value2 = value2
+            }
+            public init(from decoder: any Swift.Decoder) throws {
+                self.value1 = try decoder.decodeFromSingleValueContainer()
+                self.value2 = try decoder.decodeFromSingleValueContainer()
+            }
+            public func encode(to encoder: any Swift.Encoder) throws {
+                try encoder.encodeToSingleValueContainer(self.value1)
+            }
+        }
+        /// Kind of pet
+        ///
+        /// - Remark: Generated from `#/components/schemas/PetKind`.
+        @frozen public enum PetKind: String, Codable, Hashable, Sendable, CaseIterable {
+            case cat = "cat"
+            case dog = "dog"
+            case elephant = "ELEPHANT"
+            case bigElephant1 = "BIG_ELEPHANT_1"
+            case _dollar_nake = "$nake"
+            case _public = "public"
+        }
+        /// - Remark: Generated from `#/components/schemas/CreatePetRequest`.
+        public struct CreatePetRequest: Codable, Hashable, Sendable {
+            /// - Remark: Generated from `#/components/schemas/CreatePetRequest/name`.
+            public var name: Swift.String
+            /// - Remark: Generated from `#/components/schemas/CreatePetRequest/kind`.
+            public var kind: Components.Schemas.PetKind?
+            /// - Remark: Generated from `#/components/schemas/CreatePetRequest/tag`.
+            public var tag: Swift.String?
+            /// - Remark: Generated from `#/components/schemas/CreatePetRequest/genome`.
+            public var genome: OpenAPIRuntime.Base64EncodedData?
+            /// Creates a new `CreatePetRequest`.
+            ///
+            /// - Parameters:
+            ///   - name:
+            ///   - kind:
+            ///   - tag:
+            ///   - genome:
+            public init(
+                name: Swift.String,
+                kind: Components.Schemas.PetKind? = nil,
+                tag: Swift.String? = nil,
+                genome: OpenAPIRuntime.Base64EncodedData? = nil
+            ) {
+                self.name = name
+                self.kind = kind
+                self.tag = tag
+                self.genome = genome
+            }
+            public enum CodingKeys: String, CodingKey {
+                case name
+                case kind
+                case tag
+                case genome
+            }
+        }
+        /// - Remark: Generated from `#/components/schemas/Pets`.
+        public typealias Pets = [Components.Schemas.Pet]
+        /// - Remark: Generated from `#/components/schemas/Error`.
+        public struct _Error: Codable, Hashable, Sendable {
+            /// - Remark: Generated from `#/components/schemas/Error/code`.
+            public var code: Swift.Int32
+            /// - Remark: Generated from `#/components/schemas/Error/me$sage`.
+            public var me_dollar_sage: Swift.String
+            /// Extra information about the error.
+            ///
+            /// - Remark: Generated from `#/components/schemas/Error/extraInfo`.
+            public var extraInfo: Components.Schemas.ExtraInfo?
+            /// Custom user-provided key-value pairs.
+            ///
+            /// - Remark: Generated from `#/components/schemas/Error/userData`.
+            public var userData: OpenAPIRuntime.OpenAPIObjectContainer?
+            /// Creates a new `_Error`.
+            ///
+            /// - Parameters:
+            ///   - code:
+            ///   - me_dollar_sage:
+            ///   - extraInfo: Extra information about the error.
+            ///   - userData: Custom user-provided key-value pairs.
+            public init(
+                code: Swift.Int32,
+                me_dollar_sage: Swift.String,
+                extraInfo: Components.Schemas.ExtraInfo? = nil,
+                userData: OpenAPIRuntime.OpenAPIObjectContainer? = nil
+            ) {
+                self.code = code
+                self.me_dollar_sage = me_dollar_sage
+                self.extraInfo = extraInfo
+                self.userData = userData
+            }
+            public enum CodingKeys: String, CodingKey {
+                case code
+                case me_dollar_sage = "me$sage"
+                case extraInfo
+                case userData
+            }
+        }
+        /// - Remark: Generated from `#/components/schemas/PetFeeding`.
+        public struct PetFeeding: Codable, Hashable, Sendable {
+            /// - Remark: Generated from `#/components/schemas/PetFeeding/schedule`.
+            @frozen public enum SchedulePayload: String, Codable, Hashable, Sendable, CaseIterable {
+                case hourly = "hourly"
+                case daily = "daily"
+                case weekly = "weekly"
+            }
+            /// - Remark: Generated from `#/components/schemas/PetFeeding/schedule`.
+            public var schedule: Components.Schemas.PetFeeding.SchedulePayload?
+            /// Creates a new `PetFeeding`.
+            ///
+            /// - Parameters:
+            ///   - schedule:
+            public init(schedule: Components.Schemas.PetFeeding.SchedulePayload? = nil) {
+                self.schedule = schedule
+            }
+            public enum CodingKeys: String, CodingKey {
+                case schedule
+            }
+        }
+        /// - Remark: Generated from `#/components/schemas/DOB`.
+        public typealias Dob = Foundation.Date
+        /// - Remark: Generated from `#/components/schemas/ExtraInfo`.
+        public typealias ExtraInfo = Swift.String
+        /// - Remark: Generated from `#/components/schemas/NoAdditionalProperties`.
+        public struct NoAdditionalProperties: Codable, Hashable, Sendable {
+            /// - Remark: Generated from `#/components/schemas/NoAdditionalProperties/foo`.
+            public var foo: Swift.String?
+            /// Creates a new `NoAdditionalProperties`.
+            ///
+            /// - Parameters:
+            ///   - foo:
+            public init(foo: Swift.String? = nil) {
+                self.foo = foo
+            }
+            public enum CodingKeys: String, CodingKey {
+                case foo
+            }
+            public init(from decoder: any Swift.Decoder) throws {
+                let container = try decoder.container(keyedBy: CodingKeys.self)
+                self.foo = try container.decodeIfPresent(
+                    Swift.String.self,
+                    forKey: .foo
+                )
+                try decoder.ensureNoAdditionalProperties(knownKeys: [
+                    "foo"
+                ])
+            }
+        }
+        /// - Remark: Generated from `#/components/schemas/AnyAdditionalProperties`.
+        public struct AnyAdditionalProperties: Codable, Hashable, Sendable {
+            /// - Remark: Generated from `#/components/schemas/AnyAdditionalProperties/foo`.
+            public var foo: Swift.String?
+            /// A container of undocumented properties.
+            public var additionalProperties: OpenAPIRuntime.OpenAPIObjectContainer
+            /// Creates a new `AnyAdditionalProperties`.
+            ///
+            /// - Parameters:
+            ///   - foo:
+            ///   - additionalProperties: A container of undocumented properties.
+            public init(
+                foo: Swift.String? = nil,
+                additionalProperties: OpenAPIRuntime.OpenAPIObjectContainer = .init()
+            ) {
+                self.foo = foo
+                self.additionalProperties = additionalProperties
+            }
+            public enum CodingKeys: String, CodingKey {
+                case foo
+            }
+            public init(from decoder: any Swift.Decoder) throws {
+                let container = try decoder.container(keyedBy: CodingKeys.self)
+                self.foo = try container.decodeIfPresent(
+                    Swift.String.self,
+                    forKey: .foo
+                )
+                additionalProperties = try decoder.decodeAdditionalProperties(knownKeys: [
+                    "foo"
+                ])
+            }
+            public func encode(to encoder: any Swift.Encoder) throws {
+                var container = encoder.container(keyedBy: CodingKeys.self)
+                try container.encodeIfPresent(
+                    self.foo,
+                    forKey: .foo
+                )
+                try encoder.encodeAdditionalProperties(additionalProperties)
+            }
+        }
+        /// - Remark: Generated from `#/components/schemas/TypedAdditionalProperties`.
+        public struct TypedAdditionalProperties: Codable, Hashable, Sendable {
+            /// - Remark: Generated from `#/components/schemas/TypedAdditionalProperties/foo`.
+            public var foo: Swift.String?
+            /// A container of undocumented properties.
+            public var additionalProperties: [String: Swift.Int]
+            /// Creates a new `TypedAdditionalProperties`.
+            ///
+            /// - Parameters:
+            ///   - foo:
+            ///   - additionalProperties: A container of undocumented properties.
+            public init(
+                foo: Swift.String? = nil,
+                additionalProperties: [String: Swift.Int] = .init()
+            ) {
+                self.foo = foo
+                self.additionalProperties = additionalProperties
+            }
+            public enum CodingKeys: String, CodingKey {
+                case foo
+            }
+            public init(from decoder: any Swift.Decoder) throws {
+                let container = try decoder.container(keyedBy: CodingKeys.self)
+                self.foo = try container.decodeIfPresent(
+                    Swift.String.self,
+                    forKey: .foo
+                )
+                additionalProperties = try decoder.decodeAdditionalProperties(knownKeys: [
+                    "foo"
+                ])
+            }
+            public func encode(to encoder: any Swift.Encoder) throws {
+                var container = encoder.container(keyedBy: CodingKeys.self)
+                try container.encodeIfPresent(
+                    self.foo,
+                    forKey: .foo
+                )
+                try encoder.encodeAdditionalProperties(additionalProperties)
+            }
+        }
+        /// - Remark: Generated from `#/components/schemas/ObjectWithOptionalNullableArrayOfNullableItems`.
+        public struct ObjectWithOptionalNullableArrayOfNullableItems: Codable, Hashable, Sendable {
+            /// - Remark: Generated from `#/components/schemas/ObjectWithOptionalNullableArrayOfNullableItems/foo`.
+            public var foo: [Swift.String?]?
+            /// Creates a new `ObjectWithOptionalNullableArrayOfNullableItems`.
+            ///
+            /// - Parameters:
+            ///   - foo:
+            public init(foo: [Swift.String?]? = nil) {
+                self.foo = foo
+            }
+            public enum CodingKeys: String, CodingKey {
+                case foo
+            }
+        }
+        /// - Remark: Generated from `#/components/schemas/TypedAdditionalPropertiesWithPropertyNamedContainer`.
+        public struct TypedAdditionalPropertiesWithPropertyNamedContainer: Codable, Hashable, Sendable {
+            /// - Remark: Generated from `#/components/schemas/TypedAdditionalPropertiesWithPropertyNamedContainer/container`.
+            public var container: Swift.String?
+            /// A container of undocumented properties.
+            public var additionalProperties: [String: Swift.Int]
+            /// Creates a new `TypedAdditionalPropertiesWithPropertyNamedContainer`.
+            ///
+            /// - Parameters:
+            ///   - container:
+            ///   - additionalProperties: A container of undocumented properties.
+            public init(
+                container: Swift.String? = nil,
+                additionalProperties: [String: Swift.Int] = .init()
+            ) {
+                self.container = container
+                self.additionalProperties = additionalProperties
+            }
+            public enum CodingKeys: String, CodingKey {
+                case container
+            }
+            public init(from decoder: any Swift.Decoder) throws {
+                let container = try decoder.container(keyedBy: CodingKeys.self)
+                self.container = try container.decodeIfPresent(
+                    Swift.String.self,
+                    forKey: .container
+                )
+                additionalProperties = try decoder.decodeAdditionalProperties(knownKeys: [
+                    "container"
+                ])
+            }
+            public func encode(to encoder: any Swift.Encoder) throws {
+                var container = encoder.container(keyedBy: CodingKeys.self)
+                try container.encodeIfPresent(
+                    self.container,
+                    forKey: .container
+                )
+                try encoder.encodeAdditionalProperties(additionalProperties)
+            }
+        }
+        /// - Remark: Generated from `#/components/schemas/CodeError`.
+        public struct CodeError: Codable, Hashable, Sendable {
+            /// - Remark: Generated from `#/components/schemas/CodeError/code`.
+            public var code: Swift.Int
+            /// Creates a new `CodeError`.
+            ///
+            /// - Parameters:
+            ///   - code:
+            public init(code: Swift.Int) {
+                self.code = code
+            }
+            public enum CodingKeys: String, CodingKey {
+                case code
+            }
+        }
+        /// - Remark: Generated from `#/components/schemas/AllOfObjects`.
+        public struct AllOfObjects: Codable, Hashable, Sendable {
+            /// - Remark: Generated from `#/components/schemas/AllOfObjects/value1`.
+            public struct Value1Payload: Codable, Hashable, Sendable {
+                /// - Remark: Generated from `#/components/schemas/AllOfObjects/value1/message`.
+                public var message: Swift.String
+                /// Creates a new `Value1Payload`.
+                ///
+                /// - Parameters:
+                ///   - message:
+                public init(message: Swift.String) {
+                    self.message = message
+                }
+                public enum CodingKeys: String, CodingKey {
+                    case message
+                }
+            }
+            /// - Remark: Generated from `#/components/schemas/AllOfObjects/value1`.
+            public var value1: Components.Schemas.AllOfObjects.Value1Payload
+            /// - Remark: Generated from `#/components/schemas/AllOfObjects/value2`.
+            public var value2: Components.Schemas.CodeError
+            /// Creates a new `AllOfObjects`.
+            ///
+            /// - Parameters:
+            ///   - value1:
+            ///   - value2:
+            public init(
+                value1: Components.Schemas.AllOfObjects.Value1Payload,
+                value2: Components.Schemas.CodeError
+            ) {
+                self.value1 = value1
+                self.value2 = value2
+            }
+            public init(from decoder: any Swift.Decoder) throws {
+                self.value1 = try .init(from: decoder)
+                self.value2 = try .init(from: decoder)
+            }
+            public func encode(to encoder: any Swift.Encoder) throws {
+                try self.value1.encode(to: encoder)
+                try self.value2.encode(to: encoder)
+            }
+        }
+        /// - Remark: Generated from `#/components/schemas/AnyOfObjects`.
+        public struct AnyOfObjects: Codable, Hashable, Sendable {
+            /// - Remark: Generated from `#/components/schemas/AnyOfObjects/value1`.
+            public struct Value1Payload: Codable, Hashable, Sendable {
+                /// - Remark: Generated from `#/components/schemas/AnyOfObjects/value1/message`.
+                public var message: Swift.String
+                /// Creates a new `Value1Payload`.
+                ///
+                /// - Parameters:
+                ///   - message:
+                public init(message: Swift.String) {
+                    self.message = message
+                }
+                public enum CodingKeys: String, CodingKey {
+                    case message
+                }
+            }
+            /// - Remark: Generated from `#/components/schemas/AnyOfObjects/value1`.
+            public var value1: Components.Schemas.AnyOfObjects.Value1Payload?
+            /// - Remark: Generated from `#/components/schemas/AnyOfObjects/value2`.
+            public var value2: Components.Schemas.CodeError?
+            /// Creates a new `AnyOfObjects`.
+            ///
+            /// - Parameters:
+            ///   - value1:
+            ///   - value2:
+            public init(
+                value1: Components.Schemas.AnyOfObjects.Value1Payload? = nil,
+                value2: Components.Schemas.CodeError? = nil
+            ) {
+                self.value1 = value1
+                self.value2 = value2
+            }
+            public init(from decoder: any Swift.Decoder) throws {
+                var errors: [any Swift.Error] = []
+                do {
+                    self.value1 = try .init(from: decoder)
+                } catch {
+                    errors.append(error)
+                }
+                do {
+                    self.value2 = try .init(from: decoder)
+                } catch {
+                    errors.append(error)
+                }
+                try Swift.DecodingError.verifyAtLeastOneSchemaIsNotNil(
+                    [
+                        self.value1,
+                        self.value2
+                    ],
+                    type: Self.self,
+                    codingPath: decoder.codingPath,
+                    errors: errors
+                )
+            }
+            public func encode(to encoder: any Swift.Encoder) throws {
+                try self.value1?.encode(to: encoder)
+                try self.value2?.encode(to: encoder)
+            }
+        }
+        /// - Remark: Generated from `#/components/schemas/OneOfAny`.
+        @frozen public enum OneOfAny: Codable, Hashable, Sendable {
+            /// - Remark: Generated from `#/components/schemas/OneOfAny/case1`.
+            case case1(Swift.String)
+            /// - Remark: Generated from `#/components/schemas/OneOfAny/case2`.
+            case case2(Swift.Int)
+            /// - Remark: Generated from `#/components/schemas/OneOfAny/case3`.
+            case CodeError(Components.Schemas.CodeError)
+            /// - Remark: Generated from `#/components/schemas/OneOfAny/case4`.
+            public struct Case4Payload: Codable, Hashable, Sendable {
+                /// - Remark: Generated from `#/components/schemas/OneOfAny/case4/message`.
+                public var message: Swift.String
+                /// Creates a new `Case4Payload`.
+                ///
+                /// - Parameters:
+                ///   - message:
+                public init(message: Swift.String) {
+                    self.message = message
+                }
+                public enum CodingKeys: String, CodingKey {
+                    case message
+                }
+            }
+            /// - Remark: Generated from `#/components/schemas/OneOfAny/case4`.
+            case case4(Components.Schemas.OneOfAny.Case4Payload)
+            public init(from decoder: any Swift.Decoder) throws {
+                var errors: [any Swift.Error] = []
+                do {
+                    self = .case1(try decoder.decodeFromSingleValueContainer())
+                    return
+                } catch {
+                    errors.append(error)
+                }
+                do {
+                    self = .case2(try decoder.decodeFromSingleValueContainer())
+                    return
+                } catch {
+                    errors.append(error)
+                }
+                do {
+                    self = .CodeError(try .init(from: decoder))
+                    return
+                } catch {
+                    errors.append(error)
+                }
+                do {
+                    self = .case4(try .init(from: decoder))
+                    return
+                } catch {
+                    errors.append(error)
+                }
+                throw Swift.DecodingError.failedToDecodeOneOfSchema(
+                    type: Self.self,
+                    codingPath: decoder.codingPath,
+                    errors: errors
+                )
+            }
+            public func encode(to encoder: any Swift.Encoder) throws {
+                switch self {
+                case let .case1(value):
+                    try encoder.encodeToSingleValueContainer(value)
+                case let .case2(value):
+                    try encoder.encodeToSingleValueContainer(value)
+                case let .CodeError(value):
+                    try value.encode(to: encoder)
+                case let .case4(value):
+                    try value.encode(to: encoder)
+                }
+            }
+        }
+        /// - Remark: Generated from `#/components/schemas/PetExercise`.
+        public struct PetExercise: Codable, Hashable, Sendable {
+            /// - Remark: Generated from `#/components/schemas/PetExercise/kind`.
+            public var kind: Swift.String
+            /// Creates a new `PetExercise`.
+            ///
+            /// - Parameters:
+            ///   - kind:
+            public init(kind: Swift.String) {
+                self.kind = kind
+            }
+            public enum CodingKeys: String, CodingKey {
+                case kind
+            }
+        }
+        /// - Remark: Generated from `#/components/schemas/Walk`.
+        public struct Walk: Codable, Hashable, Sendable {
+            /// - Remark: Generated from `#/components/schemas/Walk/kind`.
+            public var kind: Swift.String
+            /// - Remark: Generated from `#/components/schemas/Walk/length`.
+            public var length: Swift.Int
+            /// Creates a new `Walk`.
+            ///
+            /// - Parameters:
+            ///   - kind:
+            ///   - length:
+            public init(
+                kind: Swift.String,
+                length: Swift.Int
+            ) {
+                self.kind = kind
+                self.length = length
+            }
+            public enum CodingKeys: String, CodingKey {
+                case kind
+                case length
+            }
+        }
+        /// - Remark: Generated from `#/components/schemas/MessagedExercise`.
+        public struct MessagedExercise: Codable, Hashable, Sendable {
+            /// - Remark: Generated from `#/components/schemas/MessagedExercise/value1`.
+            public var value1: Components.Schemas.PetExercise
+            /// - Remark: Generated from `#/components/schemas/MessagedExercise/value2`.
+            public struct Value2Payload: Codable, Hashable, Sendable {
+                /// - Remark: Generated from `#/components/schemas/MessagedExercise/value2/message`.
+                public var message: Swift.String
+                /// Creates a new `Value2Payload`.
+                ///
+                /// - Parameters:
+                ///   - message:
+                public init(message: Swift.String) {
+                    self.message = message
+                }
+                public enum CodingKeys: String, CodingKey {
+                    case message
+                }
+            }
+            /// - Remark: Generated from `#/components/schemas/MessagedExercise/value2`.
+            public var value2: Components.Schemas.MessagedExercise.Value2Payload
+            /// Creates a new `MessagedExercise`.
+            ///
+            /// - Parameters:
+            ///   - value1:
+            ///   - value2:
+            public init(
+                value1: Components.Schemas.PetExercise,
+                value2: Components.Schemas.MessagedExercise.Value2Payload
+            ) {
+                self.value1 = value1
+                self.value2 = value2
+            }
+            public init(from decoder: any Swift.Decoder) throws {
+                self.value1 = try .init(from: decoder)
+                self.value2 = try .init(from: decoder)
+            }
+            public func encode(to encoder: any Swift.Encoder) throws {
+                try self.value1.encode(to: encoder)
+                try self.value2.encode(to: encoder)
+            }
+        }
+        /// - Remark: Generated from `#/components/schemas/OneOfObjectsWithDiscriminator`.
+        @frozen public enum OneOfObjectsWithDiscriminator: Codable, Hashable, Sendable {
+            /// - Remark: Generated from `#/components/schemas/OneOfObjectsWithDiscriminator/Walk`.
+            case walk(Components.Schemas.Walk)
+            /// - Remark: Generated from `#/components/schemas/OneOfObjectsWithDiscriminator/MessagedExercise`.
+            case messagedExercise(Components.Schemas.MessagedExercise)
+            public enum CodingKeys: String, CodingKey {
+                case kind
+            }
+            public init(from decoder: any Swift.Decoder) throws {
+                let container = try decoder.container(keyedBy: CodingKeys.self)
+                let discriminator = try container.decode(
+                    Swift.String.self,
+                    forKey: .kind
+                )
+                switch discriminator {
+                case "Walk", "#/components/schemas/Walk":
+                    self = .walk(try .init(from: decoder))
+                case "MessagedExercise", "#/components/schemas/MessagedExercise":
+                    self = .messagedExercise(try .init(from: decoder))
+                default:
+                    throw Swift.DecodingError.unknownOneOfDiscriminator(
+                        discriminatorKey: CodingKeys.kind,
+                        discriminatorValue: discriminator,
+                        codingPath: decoder.codingPath
+                    )
+                }
+            }
+            public func encode(to encoder: any Swift.Encoder) throws {
+                switch self {
+                case let .walk(value):
+                    try value.encode(to: encoder)
+                case let .messagedExercise(value):
+                    try value.encode(to: encoder)
+                }
+            }
+        }
+        /// - Remark: Generated from `#/components/schemas/PetStats`.
+        public struct PetStats: Codable, Hashable, Sendable {
+            /// - Remark: Generated from `#/components/schemas/PetStats/count`.
+            public var count: Swift.Int
+            /// Creates a new `PetStats`.
+            ///
+            /// - Parameters:
+            ///   - count:
+            public init(count: Swift.Int) {
+                self.count = count
+            }
+            public enum CodingKeys: String, CodingKey {
+                case count
+            }
+        }
+        /// - Remark: Generated from `#/components/schemas/RecursivePet`.
+        public struct RecursivePet: Codable, Hashable, Sendable {
+            /// - Remark: Generated from `#/components/schemas/RecursivePet/name`.
+            public var name: Swift.String {
+                get  {
+                    self.storage.value.name
+                }
+                _modify {
+                    yield &self.storage.value.name
+                }
+            }
+            /// - Remark: Generated from `#/components/schemas/RecursivePet/parent`.
+            public var parent: Components.Schemas.RecursivePet? {
+                get  {
+                    self.storage.value.parent
+                }
+                _modify {
+                    yield &self.storage.value.parent
+                }
+            }
+            /// Creates a new `RecursivePet`.
+            ///
+            /// - Parameters:
+            ///   - name:
+            ///   - parent:
+            public init(
+                name: Swift.String,
+                parent: Components.Schemas.RecursivePet? = nil
+            ) {
+                self.storage = .init(value: .init(
+                    name: name,
+                    parent: parent
+                ))
+            }
+            public enum CodingKeys: String, CodingKey {
+                case name
+                case parent
+            }
+            public init(from decoder: any Swift.Decoder) throws {
+                self.storage = try .init(from: decoder)
+            }
+            public func encode(to encoder: any Swift.Encoder) throws {
+                try self.storage.encode(to: encoder)
+            }
+            /// Internal reference storage to allow type recursion.
+            private var storage: OpenAPIRuntime.CopyOnWriteBox<Storage>
+            private struct Storage: Codable, Hashable, Sendable {
+                /// - Remark: Generated from `#/components/schemas/RecursivePet/name`.
+                var name: Swift.String
+                /// - Remark: Generated from `#/components/schemas/RecursivePet/parent`.
+                var parent: Components.Schemas.RecursivePet?
+                init(
+                    name: Swift.String,
+                    parent: Components.Schemas.RecursivePet? = nil
+                ) {
+                    self.name = name
+                    self.parent = parent
+                }
+                typealias CodingKeys = Components.Schemas.RecursivePet.CodingKeys
+            }
+        }
+        /// - Remark: Generated from `#/components/schemas/RecursivePetNested`.
+        public struct RecursivePetNested: Codable, Hashable, Sendable {
+            /// - Remark: Generated from `#/components/schemas/RecursivePetNested/name`.
+            public var name: Swift.String {
+                get  {
+                    self.storage.value.name
+                }
+                _modify {
+                    yield &self.storage.value.name
+                }
+            }
+            /// - Remark: Generated from `#/components/schemas/RecursivePetNested/parent`.
+            public struct ParentPayload: Codable, Hashable, Sendable {
+                /// - Remark: Generated from `#/components/schemas/RecursivePetNested/parent/nested`.
+                public var nested: Components.Schemas.RecursivePetNested
+                /// Creates a new `ParentPayload`.
+                ///
+                /// - Parameters:
+                ///   - nested:
+                public init(nested: Components.Schemas.RecursivePetNested) {
+                    self.nested = nested
+                }
+                public enum CodingKeys: String, CodingKey {
+                    case nested
+                }
+            }
+            /// - Remark: Generated from `#/components/schemas/RecursivePetNested/parent`.
+            public var parent: Components.Schemas.RecursivePetNested.ParentPayload? {
+                get  {
+                    self.storage.value.parent
+                }
+                _modify {
+                    yield &self.storage.value.parent
+                }
+            }
+            /// Creates a new `RecursivePetNested`.
+            ///
+            /// - Parameters:
+            ///   - name:
+            ///   - parent:
+            public init(
+                name: Swift.String,
+                parent: Components.Schemas.RecursivePetNested.ParentPayload? = nil
+            ) {
+                self.storage = .init(value: .init(
+                    name: name,
+                    parent: parent
+                ))
+            }
+            public enum CodingKeys: String, CodingKey {
+                case name
+                case parent
+            }
+            public init(from decoder: any Swift.Decoder) throws {
+                self.storage = try .init(from: decoder)
+            }
+            public func encode(to encoder: any Swift.Encoder) throws {
+                try self.storage.encode(to: encoder)
+            }
+            /// Internal reference storage to allow type recursion.
+            private var storage: OpenAPIRuntime.CopyOnWriteBox<Storage>
+            private struct Storage: Codable, Hashable, Sendable {
+                /// - Remark: Generated from `#/components/schemas/RecursivePetNested/name`.
+                var name: Swift.String
+                /// - Remark: Generated from `#/components/schemas/RecursivePetNested/parent`.
+                struct ParentPayload: Codable, Hashable, Sendable {
+                    /// - Remark: Generated from `#/components/schemas/RecursivePetNested/parent/nested`.
+                    public var nested: Components.Schemas.RecursivePetNested
+                    /// Creates a new `ParentPayload`.
+                    ///
+                    /// - Parameters:
+                    ///   - nested:
+                    public init(nested: Components.Schemas.RecursivePetNested) {
+                        self.nested = nested
+                    }
+                    public enum CodingKeys: String, CodingKey {
+                        case nested
+                    }
+                }
+                /// - Remark: Generated from `#/components/schemas/RecursivePetNested/parent`.
+                var parent: Components.Schemas.RecursivePetNested.ParentPayload?
+                init(
+                    name: Swift.String,
+                    parent: Components.Schemas.RecursivePetNested.ParentPayload? = nil
+                ) {
+                    self.name = name
+                    self.parent = parent
+                }
+                typealias CodingKeys = Components.Schemas.RecursivePetNested.CodingKeys
+            }
+        }
+        /// - Remark: Generated from `#/components/schemas/RecursivePetOneOfFirst`.
+        public struct RecursivePetOneOfFirst: Codable, Hashable, Sendable {
+            /// - Remark: Generated from `#/components/schemas/RecursivePetOneOfFirst/value1`.
+            public var value1: Components.Schemas.RecursivePetOneOf {
+                get  {
+                    self.storage.value.value1
+                }
+                _modify {
+                    yield &self.storage.value.value1
+                }
+            }
+            /// - Remark: Generated from `#/components/schemas/RecursivePetOneOfFirst/value2`.
+            public struct Value2Payload: Codable, Hashable, Sendable {
+                /// - Remark: Generated from `#/components/schemas/RecursivePetOneOfFirst/value2/type`.
+                public var _type: Swift.String
+                /// Creates a new `Value2Payload`.
+                ///
+                /// - Parameters:
+                ///   - _type:
+                public init(_type: Swift.String) {
+                    self._type = _type
+                }
+                public enum CodingKeys: String, CodingKey {
+                    case _type = "type"
+                }
+            }
+            /// - Remark: Generated from `#/components/schemas/RecursivePetOneOfFirst/value2`.
+            public var value2: Components.Schemas.RecursivePetOneOfFirst.Value2Payload {
+                get  {
+                    self.storage.value.value2
+                }
+                _modify {
+                    yield &self.storage.value.value2
+                }
+            }
+            /// Creates a new `RecursivePetOneOfFirst`.
+            ///
+            /// - Parameters:
+            ///   - value1:
+            ///   - value2:
+            public init(
+                value1: Components.Schemas.RecursivePetOneOf,
+                value2: Components.Schemas.RecursivePetOneOfFirst.Value2Payload
+            ) {
+                self.storage = .init(value: .init(
+                    value1: value1,
+                    value2: value2
+                ))
+            }
+            public init(from decoder: any Swift.Decoder) throws {
+                self.storage = try .init(from: decoder)
+            }
+            public func encode(to encoder: any Swift.Encoder) throws {
+                try self.storage.encode(to: encoder)
+            }
+            /// Internal reference storage to allow type recursion.
+            private var storage: OpenAPIRuntime.CopyOnWriteBox<Storage>
+            private struct Storage: Codable, Hashable, Sendable {
+                /// - Remark: Generated from `#/components/schemas/RecursivePetOneOfFirst/value1`.
+                var value1: Components.Schemas.RecursivePetOneOf
+                /// - Remark: Generated from `#/components/schemas/RecursivePetOneOfFirst/value2`.
+                struct Value2Payload: Codable, Hashable, Sendable {
+                    /// - Remark: Generated from `#/components/schemas/RecursivePetOneOfFirst/value2/type`.
+                    public var _type: Swift.String
+                    /// Creates a new `Value2Payload`.
+                    ///
+                    /// - Parameters:
+                    ///   - _type:
+                    public init(_type: Swift.String) {
+                        self._type = _type
+                    }
+                    public enum CodingKeys: String, CodingKey {
+                        case _type = "type"
+                    }
+                }
+                /// - Remark: Generated from `#/components/schemas/RecursivePetOneOfFirst/value2`.
+                var value2: Components.Schemas.RecursivePetOneOfFirst.Value2Payload
+                init(
+                    value1: Components.Schemas.RecursivePetOneOf,
+                    value2: Components.Schemas.RecursivePetOneOfFirst.Value2Payload
+                ) {
+                    self.value1 = value1
+                    self.value2 = value2
+                }
+                init(from decoder: any Swift.Decoder) throws {
+                    self.value1 = try .init(from: decoder)
+                    self.value2 = try .init(from: decoder)
+                }
+                func encode(to encoder: any Swift.Encoder) throws {
+                    try self.value1.encode(to: encoder)
+                    try self.value2.encode(to: encoder)
+                }
+            }
+        }
+        /// - Remark: Generated from `#/components/schemas/RecursivePetOneOfSecond`.
+        public struct RecursivePetOneOfSecond: Codable, Hashable, Sendable {
+            /// - Remark: Generated from `#/components/schemas/RecursivePetOneOfSecond/value1`.
+            public var value1: Components.Schemas.Pet
+            /// - Remark: Generated from `#/components/schemas/RecursivePetOneOfSecond/value2`.
+            public struct Value2Payload: Codable, Hashable, Sendable {
+                /// - Remark: Generated from `#/components/schemas/RecursivePetOneOfSecond/value2/type`.
+                public var _type: Swift.String
+                /// Creates a new `Value2Payload`.
+                ///
+                /// - Parameters:
+                ///   - _type:
+                public init(_type: Swift.String) {
+                    self._type = _type
+                }
+                public enum CodingKeys: String, CodingKey {
+                    case _type = "type"
+                }
+            }
+            /// - Remark: Generated from `#/components/schemas/RecursivePetOneOfSecond/value2`.
+            public var value2: Components.Schemas.RecursivePetOneOfSecond.Value2Payload
+            /// Creates a new `RecursivePetOneOfSecond`.
+            ///
+            /// - Parameters:
+            ///   - value1:
+            ///   - value2:
+            public init(
+                value1: Components.Schemas.Pet,
+                value2: Components.Schemas.RecursivePetOneOfSecond.Value2Payload
+            ) {
+                self.value1 = value1
+                self.value2 = value2
+            }
+            public init(from decoder: any Swift.Decoder) throws {
+                self.value1 = try .init(from: decoder)
+                self.value2 = try .init(from: decoder)
+            }
+            public func encode(to encoder: any Swift.Encoder) throws {
+                try self.value1.encode(to: encoder)
+                try self.value2.encode(to: encoder)
+            }
+        }
+        /// - Remark: Generated from `#/components/schemas/RecursivePetOneOf`.
+        @frozen public enum RecursivePetOneOf: Codable, Hashable, Sendable {
+            /// - Remark: Generated from `#/components/schemas/RecursivePetOneOf/RecursivePetOneOfFirst`.
+            case recursivePetOneOfFirst(Components.Schemas.RecursivePetOneOfFirst)
+            /// - Remark: Generated from `#/components/schemas/RecursivePetOneOf/RecursivePetOneOfSecond`.
+            case recursivePetOneOfSecond(Components.Schemas.RecursivePetOneOfSecond)
+            public enum CodingKeys: String, CodingKey {
+                case _type = "type"
+            }
+            public init(from decoder: any Swift.Decoder) throws {
+                let container = try decoder.container(keyedBy: CodingKeys.self)
+                let discriminator = try container.decode(
+                    Swift.String.self,
+                    forKey: ._type
+                )
+                switch discriminator {
+                case "RecursivePetOneOfFirst", "#/components/schemas/RecursivePetOneOfFirst":
+                    self = .recursivePetOneOfFirst(try .init(from: decoder))
+                case "RecursivePetOneOfSecond", "#/components/schemas/RecursivePetOneOfSecond":
+                    self = .recursivePetOneOfSecond(try .init(from: decoder))
+                default:
+                    throw Swift.DecodingError.unknownOneOfDiscriminator(
+                        discriminatorKey: CodingKeys._type,
+                        discriminatorValue: discriminator,
+                        codingPath: decoder.codingPath
+                    )
+                }
+            }
+            public func encode(to encoder: any Swift.Encoder) throws {
+                switch self {
+                case let .recursivePetOneOfFirst(value):
+                    try value.encode(to: encoder)
+                case let .recursivePetOneOfSecond(value):
+                    try value.encode(to: encoder)
+                }
+            }
+        }
+        /// - Remark: Generated from `#/components/schemas/RecursivePetAnyOf`.
+        public struct RecursivePetAnyOf: Codable, Hashable, Sendable {
+            /// - Remark: Generated from `#/components/schemas/RecursivePetAnyOf/value1`.
+            public var value1: Components.Schemas.RecursivePetAnyOf? {
+                get  {
+                    self.storage.value.value1
+                }
+                _modify {
+                    yield &self.storage.value.value1
+                }
+            }
+            /// - Remark: Generated from `#/components/schemas/RecursivePetAnyOf/value2`.
+            public var value2: Swift.String? {
+                get  {
+                    self.storage.value.value2
+                }
+                _modify {
+                    yield &self.storage.value.value2
+                }
+            }
+            /// Creates a new `RecursivePetAnyOf`.
+            ///
+            /// - Parameters:
+            ///   - value1:
+            ///   - value2:
+            public init(
+                value1: Components.Schemas.RecursivePetAnyOf? = nil,
+                value2: Swift.String? = nil
+            ) {
+                self.storage = .init(value: .init(
+                    value1: value1,
+                    value2: value2
+                ))
+            }
+            public init(from decoder: any Swift.Decoder) throws {
+                self.storage = try .init(from: decoder)
+            }
+            public func encode(to encoder: any Swift.Encoder) throws {
+                try self.storage.encode(to: encoder)
+            }
+            /// Internal reference storage to allow type recursion.
+            private var storage: OpenAPIRuntime.CopyOnWriteBox<Storage>
+            private struct Storage: Codable, Hashable, Sendable {
+                /// - Remark: Generated from `#/components/schemas/RecursivePetAnyOf/value1`.
+                var value1: Components.Schemas.RecursivePetAnyOf?
+                /// - Remark: Generated from `#/components/schemas/RecursivePetAnyOf/value2`.
+                var value2: Swift.String?
+                init(
+                    value1: Components.Schemas.RecursivePetAnyOf? = nil,
+                    value2: Swift.String? = nil
+                ) {
+                    self.value1 = value1
+                    self.value2 = value2
+                }
+                init(from decoder: any Swift.Decoder) throws {
+                    var errors: [any Swift.Error] = []
+                    do {
+                        self.value1 = try .init(from: decoder)
+                    } catch {
+                        errors.append(error)
+                    }
+                    do {
+                        self.value2 = try decoder.decodeFromSingleValueContainer()
+                    } catch {
+                        errors.append(error)
+                    }
+                    try Swift.DecodingError.verifyAtLeastOneSchemaIsNotNil(
+                        [
+                            self.value1,
+                            self.value2
+                        ],
+                        type: Self.self,
+                        codingPath: decoder.codingPath,
+                        errors: errors
+                    )
+                }
+                func encode(to encoder: any Swift.Encoder) throws {
+                    try encoder.encodeFirstNonNilValueToSingleValueContainer([
+                        self.value2
+                    ])
+                    try self.value1?.encode(to: encoder)
+                }
+            }
+        }
+        /// - Remark: Generated from `#/components/schemas/RecursivePetAllOf`.
+        public struct RecursivePetAllOf: Codable, Hashable, Sendable {
+            /// - Remark: Generated from `#/components/schemas/RecursivePetAllOf/value1`.
+            public struct Value1Payload: Codable, Hashable, Sendable {
+                /// - Remark: Generated from `#/components/schemas/RecursivePetAllOf/value1/parent`.
+                public var parent: Components.Schemas.RecursivePetAllOf?
+                /// Creates a new `Value1Payload`.
+                ///
+                /// - Parameters:
+                ///   - parent:
+                public init(parent: Components.Schemas.RecursivePetAllOf? = nil) {
+                    self.parent = parent
+                }
+                public enum CodingKeys: String, CodingKey {
+                    case parent
+                }
+            }
+            /// - Remark: Generated from `#/components/schemas/RecursivePetAllOf/value1`.
+            public var value1: Components.Schemas.RecursivePetAllOf.Value1Payload {
+                get  {
+                    self.storage.value.value1
+                }
+                _modify {
+                    yield &self.storage.value.value1
+                }
+            }
+            /// Creates a new `RecursivePetAllOf`.
+            ///
+            /// - Parameters:
+            ///   - value1:
+            public init(value1: Components.Schemas.RecursivePetAllOf.Value1Payload) {
+                self.storage = .init(value: .init(value1: value1))
+            }
+            public init(from decoder: any Swift.Decoder) throws {
+                self.storage = try .init(from: decoder)
+            }
+            public func encode(to encoder: any Swift.Encoder) throws {
+                try self.storage.encode(to: encoder)
+            }
+            /// Internal reference storage to allow type recursion.
+            private var storage: OpenAPIRuntime.CopyOnWriteBox<Storage>
+            private struct Storage: Codable, Hashable, Sendable {
+                /// - Remark: Generated from `#/components/schemas/RecursivePetAllOf/value1`.
+                struct Value1Payload: Codable, Hashable, Sendable {
+                    /// - Remark: Generated from `#/components/schemas/RecursivePetAllOf/value1/parent`.
+                    public var parent: Components.Schemas.RecursivePetAllOf?
+                    /// Creates a new `Value1Payload`.
+                    ///
+                    /// - Parameters:
+                    ///   - parent:
+                    public init(parent: Components.Schemas.RecursivePetAllOf? = nil) {
+                        self.parent = parent
+                    }
+                    public enum CodingKeys: String, CodingKey {
+                        case parent
+                    }
+                }
+                /// - Remark: Generated from `#/components/schemas/RecursivePetAllOf/value1`.
+                var value1: Components.Schemas.RecursivePetAllOf.Value1Payload
+                init(value1: Components.Schemas.RecursivePetAllOf.Value1Payload) {
+                    self.value1 = value1
+                }
+                init(from decoder: any Swift.Decoder) throws {
+                    self.value1 = try .init(from: decoder)
+                }
+                func encode(to encoder: any Swift.Encoder) throws {
+                    try self.value1.encode(to: encoder)
+                }
+            }
+        }
+    }
+}
