+++ Tests/OpenAPIGeneratorTests/Test_GenerateOptions.swift
@@ -52,6 +52,42 @@ final class Test_GenerateOptions: XCTestCase {
         } catch { XCTFail("Expected to throw a Diagnostic `.error`, but threw a different error: \(error)") }
     }
 
+    func testBuildPluginWritesEveryDeclaredOutputForUnrequestedModes() async throws {
+        let temporaryDirectory = FileManager.default.temporaryDirectory.appendingPathComponent(UUID().uuidString)
+        try FileManager.default.createDirectory(at: temporaryDirectory, withIntermediateDirectories: true)
+        defer { try? FileManager.default.removeItem(at: temporaryDirectory) }
+
+        let documentURL = temporaryDirectory.appendingPathComponent("openapi.yaml")
+        try Data(
+            """
+            openapi: "3.1.0"
+            info:
+              title: GreetingService
+              version: "1.0.0"
+            paths: {}
+            """.utf8
+        ).write(to: documentURL)
+
+        try await _Tool.runGenerator(
+            doc: documentURL,
+            configs: [Config(mode: .client, access: .internal, namingStrategy: .defensive)],
+            pluginSource: .build,
+            outputDirectory: temporaryDirectory,
+            isDryRun: false,
+            diagnostics: StdErrPrintingDiagnosticCollector()
+        )
+
+        for outputFileName in GeneratorMode.allOutputFileNames {
+            let outputURL = temporaryDirectory.appendingPathComponent(outputFileName)
+            XCTAssertTrue(FileManager.default.fileExists(atPath: outputURL.path), "Missing \(outputFileName)")
+        }
+        for outputFileName in GeneratorMode.types.outputFileNames + GeneratorMode.server.outputFileNames {
+            let outputURL = temporaryDirectory.appendingPathComponent(outputFileName)
+            XCTAssertEqual(try Data(contentsOf: outputURL), Data(), "Expected empty \(outputFileName)")
+        }
+        XCTAssertFalse(try Data(contentsOf: temporaryDirectory.appendingPathComponent("Client.swift")).isEmpty)
+    }
+
     /// Tests that `handleFileOperation` correctly transforms file-not-found errors into user-friendly messages.
     /// This test verifies the error handling works correctly on both macOS and Linux.
     func testHandleFileOperation_FileNotFound() throws {
@@ -142,60 +178,6 @@ final class Test_GenerateOptions: XCTestCase {
         } catch { XCTFail("Expected ArgumentParser.ValidationError, but got: \(type(of: error)) - \(error)") }
     }
 
-    func testLoadedConfigDecodesOutputOptions() throws {
-        let configURL = try makeTemporaryConfig(
-            """
-            generate:
-              - types
-            output:
-              types:
-                fileSplitting:
-                  strategy: namespace
-            """
-        )
-        let options = try _GenerateOptions.parse(["openapi.yaml", "--config", configURL.path])
-        let config = try XCTUnwrap(options.loadedConfig())
-
-        XCTAssertEqual(config.output?.types?.fileSplitting?.strategy, .namespace)
-    }
-
-    func testTypesFileSplittingOptionResolvesOutputOptions() throws {
-        let options = try _GenerateOptions.parse([
-            "openapi.yaml", "--mode", "types", "--types-file-splitting", "namespace",
-        ])
-
-        XCTAssertEqual(options.resolvedOutputOptions(nil).types?.fileSplitting?.strategy, .namespace)
-    }
-
-    func testTypesFileSplittingIsRejectedForBuildToolPlugin() async throws {
-        let configURL = try makeTemporaryConfig(
-            """
-            generate:
-              - types
-            output:
-              types:
-                fileSplitting:
-                  strategy: namespace
-            """
-        )
-        let options = try _GenerateOptions.parse(["openapi.yaml", "--config", configURL.path])
-
-        do {
-            try await options.runGenerator(
-                outputDirectory: URL(fileURLWithPath: "/tmp/generated"),
-                pluginSource: .build,
-                isDryRun: true
-            )
-            XCTFail("Expected build tool plugin invocation to reject types file splitting")
-        } catch let error as ArgumentParser.ValidationError {
-            XCTAssertTrue(
-                String(describing: error).contains("Types file splitting is not supported by the build tool plugin yet"),
-                "Unexpected error: \(error)"
-            )
-        } catch {
-            XCTFail("Expected ArgumentParser.ValidationError, but got: \(type(of: error)) - \(error)")
-        }
-    }
     #endif
 
     private func makeTemporaryConfig(_ contents: String) throws -> URL {
